move examples, also add docs link in app footer
