+++ src/example/entry.js
@@ -6,4 +6,4 @@ import './base.scss';
 
 // TODO: Problem with importing helpers for some reason... :/
 /* eslint-disable-next-line flowtype-errors/show-errors */
-import './helpers/relative-href.helper';
+import '../helpers/relative-href.helper';
