+++ src/example/index.html
@@ -5,11 +5,11 @@
 
 </head>
 <body class="flex flex-col min-h-screen">
-  @require("./static/header.html", { "page": "home" })
+  @require("../static/header.html", { "page": "home" })
 
   <div class="flex-grow px-3 py-5">
     <p>Welcome to the Demo FEND Application using Webpack to compile the HTML/JS/CSS files.</p>
   </div>
 
-  @require("./static/footer.html")
+  @require("../static/footer.html")
 </body>
