+++ src/example/list/components/my-list-item.hbs
