+++ src/example/list/index.html
@@ -5,10 +5,10 @@
 
 </head>
 <body class="flex flex-col min-h-screen">
-  @require("../static/header.html", { "page": "list" })
+  @require("../../static/header.html", { "page": "list" })
 
   <div id="page" class="flex-grow max-w-lg m-auto w-4/5">
   </div>
 
-  @require("../static/footer.html")
+  @require("../../static/footer.html")
 </body>
