+++ src/example/storage/app.controller.js
@@ -1,7 +1,7 @@
 /**
  * @flow
  */
-import LocalStorageHelper from '../helpers/local-storage.helper';
+import LocalStorageHelper from '../../helpers/local-storage.helper';
 import './app.scss';
 /**
  * Controller for the application
