+++ src/example/storage/entry.js
@@ -2,8 +2,8 @@
  * @flow
  */
 import '../base.scss';
-import '../helpers/relative-href.helper';
-import LocalStorageHelper from '../helpers/local-storage.helper';
+import '../../helpers/relative-href.helper';
+import LocalStorageHelper from '../../helpers/local-storage.helper';
 import StorageAppController from './app.controller';
 
 LocalStorageHelper.save({ key: 'test', val: 'abc' });
