+++ src/example/storage/index.html
@@ -5,7 +5,7 @@
 
 </head>
 <body class="flex flex-col min-h-screen">
-  @require("../static/header.html", { "page": "storage" })
+  @require("../../static/header.html", { "page": "storage" })
 
   <div class="flex-grow px-2 py-1">
     <div class="input">
@@ -37,5 +37,5 @@
     </div>
   </div>
 
-  @require("../static/footer.html")
+  @require("../../static/footer.html")
 </body>
