+++ src/example/todo/entry.js
@@ -3,7 +3,7 @@
  */
 import '../base.scss';
 import './todo.scss';
-import '../helpers/relative-href.helper';
+import '../../helpers/relative-href.helper';
 import TodoAppController from './app.controller.js';
 
 new TodoAppController();
