+++ src/example/todo/index.html
@@ -5,7 +5,7 @@
 
 </head>
 <body class="flex flex-col min-h-screen">
-  @require("../static/header.html", { "page": "todo" })
+  @require("../../static/header.html", { "page": "todo" })
 
   <div class="flex-grow max-w-lg m-auto w-4/5">
     <div id="add-container" class="border-b border-gray-500 px-2 py-1"></div>
@@ -16,5 +16,5 @@
     </form>
   </div>
 
-  @require("../static/footer.html")
+  @require("../../static/footer.html")
 </body>
