+++ src/static/footer.html
@@ -1,6 +1,8 @@
 <footer class="font-thin text-sm px-2 flex justify-between bg-opacity-50 bg-gray-300">
-  <div class="pt-1 pb-1 justify-start">
+  <div class="pt-1 pb-1 justify-start flex-1">
     @githubImg()
+
+    <a class="ml-1 text-xs" href="https://devlinjunker.github.io/template.webpack.fend">Docs</a>
   </div>
 
   <!--nav>
@@ -15,6 +17,6 @@
     <span class="text-xxs hidden" onclick="document.querySelector('.copyright-build .hidden').classList.remove('hidden'); this.classList.add('hidden');">
       <p>@branch() - @commit()</p>
       <p> @ @datetime() </p>
-    </span> 
+    </span>
   </div>
 </footer>
